PC-Agent/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
comtypes==1.4.10
dashscope==1.22.1
frozenlist==1.5.0
idna==3.10
MouseInfo==0.1.3
multidict==6.1.0
numpy==2.2.3
pillow==11.1.0
propcache==0.3.0
psutil==7.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
pywin32==308
pywinauto==0.6.9
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
yarl==1.18.3
openai
alibabacloud_tea_util
alibabacloud_ocr_api20210707
alibabacloud_tea_openapi
modelscope==1.15.0
